test(incremental): move window tests into existing IncrementalFetchingConfigTest, drop parallel file
Consolidated window and column-type test methods into the canonical test location
following the repo's convention (tests/ValueObject/ for value-object tests).
Removed the temporarily-created parallel test file and empty directories.
