+++ tests/ValueObject/IncrementalFetchingConfigTest.php
@@ -47,4 +47,41 @@ class IncrementalFetchingConfigTest extends TestCase
         Assert::assertSame(true, $config->hasLimit());
         Assert::assertSame(456, $config->getLimit());
     }
+
+    public function testWindowRoundTripAndColumnTypeIsImmutable(): void
+    {
+        $cfg = IncrementalFetchingConfig::fromArray([
+            'incrementalFetchingColumn' => 'ts',
+            'incrementalFetchingStart' => '10 days ago',
+        ]);
+        self::assertNotNull($cfg);
+        self::assertTrue($cfg->hasWindow());
+        self::assertSame('10 days ago', $cfg->getWindowStart());
+        self::assertNull($cfg->getWindowEnd());
+
+        $typed = $cfg->withColumnType('TIMESTAMP');
+        self::assertSame('TIMESTAMP', $typed->getColumnType());
+
+        // immutability: the original is untouched and still throws
+        $this->expectException(PropertyNotSetException::class);
+        $cfg->getColumnType();
+    }
+
+    public function testNoWindowAndEmptyStringsAreNotAWindow(): void
+    {
+        $none = IncrementalFetchingConfig::fromArray(['incrementalFetchingColumn' => 'id']);
+        self::assertNotNull($none);
+        self::assertFalse($none->hasWindow());
+        self::assertNull($none->getWindowStart());
+
+        $empty = IncrementalFetchingConfig::fromArray([
+            'incrementalFetchingColumn' => 'id',
+            'incrementalFetchingStart' => '',
+            'incrementalFetchingEnd' => '',
+        ]);
+        self::assertNotNull($empty);
+        self::assertFalse($empty->hasWindow());
+        self::assertNull($empty->getWindowStart());
+        self::assertNull($empty->getWindowEnd());
+    }
 }
